requirements.txt
watchdog==3.0.0
python-dotenv==1.0.1
